confirmation dialog for account removal

+++ lib/widgets/profile/profile_options_list.dart
@@ -5,6 +5,7 @@ import 'package:BSApp/screens/profile/added_deals_screen.dart';
 import 'package:BSApp/services/api_provider.dart';
 import 'package:BSApp/widgets/profile/profile_option_item.dart';
 import 'package:BSApp/widgets/profile/profile_options_header.dart';
+import 'package:confirm_dialog/confirm_dialog.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 
@@ -43,7 +44,17 @@ class ProfileOptionsList extends StatelessWidget {
     );
   }
 
-  void _deleteAccountFunction(BuildContext context) async {
-    await Provider.of<Auth>(context, listen: false).deleteAccount();
+  Future<bool> _deleteAccountFunction(BuildContext context) async {
+    var shouldDelete = await confirm(
+      context,
+      textOK: Text('POTWIERDZAM - Usuń konto'),
+      textCancel: Text('Anuluj'),
+      title: Text('Usuwanie konta'),
+      content: Text('Czy jesteś pewien, że chcesz usunąć konto? Nie można cofnąc tej operacj!'),
+    );
+    if (shouldDelete) {
+      await Provider.of<Auth>(context, listen: false).deleteAccount();
+    }
+    return Future.value(shouldDelete);
   }
 }
